backend/requirements.txt
APScheduler==3.10.4
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
colorama==0.4.6
dnspython==2.7.0
filelock==3.20.3
Flask==3.1.2
Flask-CORS==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==23.0.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
python-nmap==0.7.1
requests==2.32.5
requests-file==3.0.1
resend==2.7.0
shodan==1.31.0
SQLAlchemy==2.0.46
tldextract==5.3.1
typing-extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5